Sanitise use of sinks by moving ownership to the corresponding classes
Sink functions are now members of the corresponding classes, namely
`Module` (and by consequence subclasses) and `Sequence`. For modules
the abstract class defaults to `ArC1.add_to_buffer` if no other sink
has been provided. In sequences default sink is None which practically
yields control of the sink to the calling object (in this case `ArC1`).
In either way this eliminates the repeated use of `sink=None` as default
sink in the codebase simplifying things a bit.