Load firmware version during initialisation if available
Occasional features might require specific versions of firmware. In
order to properly differentiate between capabilities the firmware should
now report any version >= 9.3. To maintain compatibility with older
firmware (that do not report versions) the timeout is set to 2 and if
it's elapsed version default to None. If that happens then consumer code
should know that no features > 9.2 are available.